New changes made by shiwali New changes made by shiwali

+++ app/Repositories/FeeRepository.php
@@ -14,6 +14,7 @@ use App\Models\FeeForClass;
 use Log;
 use Session;
 
+
 class FeeRepository {
 
     /**
@@ -24,7 +25,7 @@ class FeeRepository {
     public function getAllFee()
     {
     	try {
-            return  $query = Fee::where('session_id',Session::get('session'))->with(['fee_head','fee_type','fee_frequency']) ->withTrashed()->get(); 
+            return  $query = Fee::where('session_id', Session::get('session'))->with(['fee_head', 'fee_type', 'fee_frequency'])->where('fee_head_id', '1')->withTrashed()->get(); 
         } catch(\Exception $err){
             Log::error('message error in getAllFee on FeeRepository :'. $err->getMessage());
             return back()->with('error', $err->getMessage());
@@ -47,7 +48,6 @@ class FeeRepository {
                 'fee_id'          => 0,
                 'fee_name'        => (old('fee_name')) ? old('fee_name') : '',
                 'fee_short_name'  => (old('fee_short_name')) ? old('fee_short_name') : '',
-                'fee_head_list'   => FeeHead::getAllFeeHeadForListing(),
                 'fee_head_id'     => 0,
                 'fee_type_list'   => FeeType::getAllFeeTypeForListing(),
                 'fee_type_id'     => 0,
@@ -62,6 +62,19 @@ class FeeRepository {
             return back()->with('error', $err->getMessage());
         }
     }
+
+
+    public function getFee($fee_head_id,$fee_type_id)
+    { 
+        try { 
+            $fee=Fee::where(['fee_head_id'=>$fee_head_id, 'fee_type_id'=>$fee_type_id])->pluck('fee_name', 'id');
+            return $fee;
+        } catch(\Exception $err){
+            Log::error('message error in getFee on FeeRepository :'. $err->getMessage());
+            return back()->with('error', $err->getMessage());
+        }
+    }
+
 
     public function getFeeHeadFrequency($fee_head_id)
     { 
@@ -81,6 +94,7 @@ class FeeRepository {
         }
     }
 
+
     public function getFeeFrequency($fee_setting_id)
     { 
         try { 
@@ -92,6 +106,7 @@ class FeeRepository {
         }
     }
 
+
     public function getFeeFrequencyValue($frequency_id)
     {
         try { 
@@ -102,6 +117,7 @@ class FeeRepository {
             return back()->with('error', $err->getMessage());
         }
     }
+
 
     /**
     * Method to create resource
@@ -145,6 +161,7 @@ class FeeRepository {
         }
     }
 
+
     /**
     * Method to fetch edit resource data
     * @param int $fee_id
@@ -154,7 +171,11 @@ class FeeRepository {
     {
         try {
             $fee = Fee::withTrashed()->with(['fee_head','fee_type','fee_frequency'])->where('id',$fee_id)->first(); //Fetch fee data
+
             $applied_in_month_ids = FeeMonth::where(['session_id'=>Session::get('session'),'fee_id'=>$fee_id])->pluck('month_id')->toArray();
+            
+            $frequency_id = ($fee->frequency_id == null) ? '0' : $fee->fee_frequency->id;
+
             // Create data for edit form
             $data = [
                 'action'          => route('update.fee'),
@@ -163,14 +184,13 @@ class FeeRepository {
                 'fee_id'          => $fee->id,
                 'fee_name'        => ($fee->fee_name) ? $fee->fee_name : old('fee_name'),
                 'fee_short_name'  => ($fee->fee_short_name) ? $fee->fee_short_name : old('fee_short_name'),
-                'fee_head_list'   => FeeHead::getAllFeeHeadForListing(),
                 'fee_head_id'     => $fee->fee_head->id,
                 'fee_type_list'   => FeeType::getAllFeeTypeForListing(),
                 'fee_type_id'     => $fee->fee_type->id,
-                'frequency_id'    => (old('frequency_id')) ? old('frequency_id') : $fee->fee_frequency->id,
-                'frequency_value' => $fee->fee_frequency->value,
+                'frequency_id'    => (old('frequency_id')) ? old('frequency_id') : $frequency_id,
+                'frequency_value' => (isset($fee->fee_frequency->value)) ? $fee->fee_frequency->value : '0',
                 'month_list'      => Month::getAllMonthForListing(),
-                'month_ids'       => $applied_in_month_ids,   
+                'month_ids'       => (empty($applied_in_month_ids)) ? array() : $applied_in_month_ids,   
             ];
             return $data;
         } catch(\Exception $err){ 
@@ -178,6 +198,7 @@ class FeeRepository {
             return back()->with('error', $err->getMessage());
         }
     }
+
 
     /**
     * Method to update resource
@@ -187,30 +208,41 @@ class FeeRepository {
     public function update($request)
     {
         try {
-            $old_month_ids = FeeMonth::where(['session_id'=>Session::get('session'),'fee_id'=>$request->fee_id])->pluck('month_id','month_id')->toArray(); //old month ids
-    
-            $month_ids=$request->month_ids; //current requested month ids
-            
-            $remove_month_ids=array_diff($old_month_ids,$month_ids);
-            
-            $new_month_ids=array_diff($month_ids,$old_month_ids);//month ids to be inserted
-           
-            //Remove month ids
-            $removed_month_ids = FeeMonth::where(['session_id'=>Session::get('session'),'fee_id'=>$request->fee_id])->whereIn('month_id',$remove_month_ids)->delete(); 
+            $response = 'false';
+            $fee  = Fee::findOrFail($request->fee_id); //Fetch fee data
 
-            //Add month ids
-            $new_data=array(); 
-            foreach ($new_month_ids as $key => $value) {
-                $new_record=array();
-                $new_record['session_id']=Session::get('session');
-                $new_record['fee_id']=$request->fee_id;
-                $new_record['month_id']=$value;
-                array_push($new_data,$new_record);
+            if($request->frequency_id != null)
+            {
+                $old_month_ids = FeeMonth::where(['session_id'=>Session::get('session'),'fee_id'=>$request->fee_id])->pluck('month_id','month_id')->toArray(); //old month ids
+        
+                $month_ids=$request->month_ids; //current requested month ids
+                
+                $remove_month_ids=array_diff($old_month_ids,$month_ids);
+                
+                $new_month_ids=array_diff($month_ids,$old_month_ids);//month ids to be inserted
+               
+                //Remove month ids
+                $removed_month_ids = FeeMonth::where(['session_id'=>Session::get('session'),'fee_id'=>$request->fee_id])->whereIn('month_id',$remove_month_ids)->delete(); 
+
+                //Add month ids
+                $new_data=array(); 
+                foreach ($new_month_ids as $key => $value) {
+                    $new_record=array();
+                    $new_record['session_id']=Session::get('session');
+                    $new_record['fee_id']=$request->fee_id;
+                    $new_record['month_id']=$value;
+                    array_push($new_data,$new_record);
+                }
+                $fee_months = FeeMonth::insert($new_data); 
+
+                if($removed_month_ids || $fee_months){
+                    $response = 'true';
+                }else {
+                    $response = 'false';
+                }
             }
-            $fee_months = FeeMonth::insert($new_data); 
             
             //Updating Fee data
-            $fee  = Fee::findOrFail($request->fee_id); //Fetch fee data
             $fee->fee_name              = $request->fee_name;
             $fee->fee_short_name        = $request->fee_short_name;
             $fee->fee_head_id           = $request->fee_head_id;
@@ -219,16 +251,19 @@ class FeeRepository {
             $fee->status                = 'Active';
             $fee->save(); // Update data
         
-            if($fee->wasChanged() || $removed_month_ids || $fee_months) { //Check if data was updated
-                return true;
-            } else {
-                return false;
+            if($fee->wasChanged()|| $removed_month_ids || $fee_months) {  
+            //Check if data was updated
+                $response = 'true';
+            }else {
+                $response = 'false';
             }
+            return $response;
         } catch(\Exception $err){
             Log::error('message error in update on FeeRepository :'. $err->getMessage());
             return back()->with('error', $err->getMessage());
         }
     }
+
 
     /**
     * Method to delete resource
@@ -250,6 +285,7 @@ class FeeRepository {
         }
     }
 
+
     /**
     * Method to delete resource
     * @param Illuminate\Http\Request
@@ -269,7 +305,6 @@ class FeeRepository {
             return back()->with('error', $err->getMessage());
         }
     }
-
